pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "AI-powered research agent using Google Generative AI"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.5",
    "uvicorn>=0.34.0",
    "langchain-google-genai>=2.0.8",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-data]
app = ["py.typed"]
